Updated Bluetooth-related items
Added fields for cell phone alerts and in-background/base-initiated phonebook transfers.
Added "Selectable Ringtone" as an option to the base cell ringtone picker if the user specifies that the base has a display or multiple ringtones.
The minimum value for a phonebook's required max capacity for phonebook transfers is now a global constant.
Bluetooth/phonebook fields are reset to default when they become inapplicable.