Updated settings UI
Added option to show/hide number of handsets (included and in collection) in phone list.
Phone list details can now be toggled by right-clicking/tapping and holding the phone list in addition to in settings.
Fixed an issue where the settings sheet on iOS/visionOS didn't show a title and done button.