Updated place-on-base power backup options/info
Added an option to specify whether a handset supports the feature. Even if a handset fits on a supported base, it doesn't necessarily mean the handset has the feature.
Updated warning text for when the feature can't be used on your phone.
Added an explanation as to why speakerphone is the only option available for calls on the handset placed on the base.