Added more handset options
Added an option to specify how buttons behave while charging.
Added options to specify whether a phone number can be dialed before placing the handset on the base for power backup to start a call, and whether the keypad is locked if power returns during a call on that handset.