Fixed acquisition/release years
Fixed an issue where the acquisition year can be set lower than the release year of the first-ever telephone when the release year was set to -1 (unknown).
Lowered minimum release year to 1876, the year the telephone was invented.
New cordless devices now have their release year set to that of the phone.
Fixed an issue where an unknown year was shown in the phone list as -1 instead of Unknown.